Make online user

+++ app/src/main/java/dev/syorito_hatsuki/onlyone/api/user/get_users/response/User.kt
@@ -14,7 +14,7 @@ data class User(
     val gender: String,
     val id: Int,
     val me: Boolean,
-    val online: Int,
+    val online: Long,
     val sendto: String,
     val status: String,
     val text: String,
